% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_code.r
\name{create.data}
\alias{create.data}
\title{Create the data for the StEM algorithm.}
\usage{
create.data(
  A,
  B,
  x.names,
  rangeA,
  rangeB,
  pivs,
  nintervalsT = 10,
  nintervalsV = 10,
  referencePIVs,
  referenceV0 = 1
)
}
\arguments{
\item{A}{Data.frame containing file A. The time of measurement must be registered as the variable "U" (see the example data).}

\item{B}{Data.frame containing file B. The time of event must be registered as the variable "V" (see the example data).}

\item{x.names}{Vector of variable names that refer to the covariates in file A.}

\item{rangeA}{Range of times at which the baseline measurements were performed. In the paper, these values are referred to as l^A and r^A.}

\item{rangeB}{Range of event times in file B. In the paper, these values are referred to as l^B and r^B.}

\item{pivs}{Vector containing the names of the partially identifying variables.}

\item{nintervalsT}{Number of intervals for the piecewise constant baseline hazard function. Default value is 10.}

\item{nintervalsV}{Number of intervals for the piecewise linear function describing the times of events without matches. Default value is 10.}

\item{referencePIVs}{Vector containing the reference values for all partially identifying variables.}

\item{referenceV0}{Index that indicates which value is the reference value for the piecewise linear function describing the times of events without matches.
The default value is the first parameter, i.e. referenceV0=1.}
}
\value{
A list of object that are used in the StEM algorithm. If necessary, the objects can be adjusted manually afterwards.
}
\description{
Create the data for the StEM algorithm.
}
\examples{
 
## See \link{StEM} function.

}
\author{
Michel H. Hof, \email{m.h.hof@amsterdamumc.nl}
}
