% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_code.r
\name{estimate.StEM}
\alias{estimate.StEM}
\title{Use the Markov chain produced by the StEM algorithm to obtain parameter estimates.}
\usage{
estimate.StEM(fit, Z0)
}
\arguments{
\item{fit}{\link{StEM} function object.}

\item{Z0}{Number of StEM iterations that are discarded as burn-in.}
}
\value{
A list containing the parameter estimates.
}
\description{
Use the Markov chain produced by the StEM algorithm to obtain parameter estimates.
}
\examples{
 
## See \link{StEM} function.
 
}
\author{
Michel H. Hof, \email{m.h.hof@amsterdamumc.nl}
}
