% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_code.r
\docType{package}
\name{RecLinkSurv-package}
\alias{RecLinkSurv}
\alias{RecLinkSurv-package}
\title{RecLinkSurv: A package for performing a time-to-event analysis with data from two files.}
\description{
The RecLinkSurv package implements the StEM-based model. For all available functions, please use library(help="RecLinkSurv").
}
\details{
import(Matrix)
}
\author{
\strong{Maintainer}: Michel H Hof \email{m.h.hof@amsterdamumc.nl}

}
