% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_code.r
\name{StEM}
\alias{StEM}
\title{Main function of the StEM-based model.}
\usage{
StEM(
  data,
  fitPrev = NULL,
  Z,
  M0,
  M1,
  trace = 1,
  ndigits = 2,
  lower = list(mu = -Inf, eta = -Inf, pi = -Inf, alpha = -Inf, beta = -Inf, gamma = -Inf),
  upper = list(mu = Inf, eta = Inf, pi = Inf, alpha = Inf, beta = Inf, gamma = Inf)
)
}
\arguments{
\item{data}{data created with \link{create.data}.}

\item{fitPrev}{Output of previous StEM run. If not previous run is available, the value should be NULL (also the default value)
Values can also be modified to change the staring values.}

\item{Z}{Number of StEM iterations; Z=Z0+Z1 if we use the notation from the paper.}

\item{M0}{Number of burn-in iterations for the Gibbs sampler.}

\item{M1}{Number of samples used in each StEM iteration.}

\item{trace}{Amount of detail that is given after performing a single StEM iteration: trace=1 gives a progressbar, trace=2 gives detailed information
onthe current parameter estimates, trace>2 also gives a plot to check whether the burn-in period was sufficient of not. Default value is 1.}

\item{ndigits}{Number of digits that is used in the output given after each StEM iteration. Default value is 2.}

\item{lower}{Lower bounds on the parameter estimates. Default values are list(mu=-Inf,eta=-Inf,pi=-Inf,alpha=-Inf,beta=-Inf,gamma=-Inf).}

\item{upper}{Upper bounds on the parameter estimates. Default values are list(mu=Inf,eta=Inf,pi=Inf,alpha=Inf,beta=Inf,gamma=Inf).}
}
\value{
A list of matrices containing the Z parameter estimates.
}
\description{
The data must be prepared with the function \link{create.data}. The model uses the same assumptions as the simulation study in the paper:
a piecewise constant baseline hazard function, a piecewise linear function for the times from events without match, and independently
distributed partially identifying variables.
}
\examples{
\dontrun{
#Example data containing 500 baseline measurements
data(A)
#Example data containing 1000 events
data(B)

head(A)
head(B)

data = create.data(A,B, x.names=c("X1","X2"), pivs=c("piv1", "piv2", "piv3", "piv4"), 
                   rangeA=c(0,1), rangeB=c(0,1), nintervalsT=5, nintervalsV=5, 
                   referencePIVs = c("a","1","1","1"), referenceV0=1)

fit1a = StEM(data, Z=250, M0=25, M1=25, trace=2, 
             lower=list(mu = -5, eta = -Inf, pi = -Inf, alpha = -Inf, beta = -Inf, gamma = -Inf))

trace.StEM(fit1a, ask=FALSE)

fit1b = estimate.StEM(fit1a, Z0=200)
fit1c = variance.StEM(fit1b, M0=100, M1=2000, data=data, Mmax=10000)
}
}
\author{
Michel H. Hof, \email{m.h.hof@amsterdamumc.nl}
}
