% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_code.r
\name{trace.StEM}
\alias{trace.StEM}
\title{Plot the Markov chain produced by the StEM algorithm.}
\usage{
trace.StEM(fit, mfrow = c(3, 2), ask = TRUE)
}
\arguments{
\item{fit}{Result from a \link{StEM} function.}

\item{mfrow}{A vector of the form c(nr, nc). Subsequent figures will be drawn in an nr-by-nc array on the device by rows, respectively.}

\item{ask}{If TRUE (and the R session is interactive) the user is asked for input before a new figure is drawn.}
}
\value{
Traceplots of the parameter estimates. The red line is the average prediction using the last 
Z iterations
}
\description{
Plot the Markov chain produced by the StEM algorithm.
}
\examples{
 
## See \link{StEM} function.

 
}
\author{
Michel H. Hof, \email{m.h.hof@amsterdamumc.nl}
}
