% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_code.r
\name{variance.StEM}
\alias{variance.StEM}
\title{Estimate the covariance matrix for a certain set of estimates. If one of the variances is negative (due to too much simulation noise), 
the algorithm continues to sample more sets of missing data until a certain maximum is reached. Note that difficulties estimating 
the covariance matrix could also indicate identifiability issues of the model. In this case, simplifying the model (e.g. lowering the number
of intervals in the piecewise constant function) might solve the problem.}
\usage{
variance.StEM(estimates, M0, M1, data, Mmax)
}
\arguments{
\item{estimates}{Parameter estimates obtained with \link{estimate.StEM}.}

\item{M0}{Number of burn-in iterations for the Gibbs sampler.}

\item{M1}{Number of terations for the Gibbs sampler. Must be larger than M0.}

\item{data}{Data created by the function \link{create.data}.}

\item{Mmax}{Maximum number of simulated sets of missing data.}
}
\value{
\item{estimates}{Estimates and variance of the estimates of the model parameters.}
\item{check}{Did the function succeed in estimating the covariance matrix: check=1 if all values on the diagonal of the covariance matrix 
are larger than zero. check=2 if Mmax has been reached but the covariance contains negative (or zero) values on the diagonal.}
\item{covar}{The complete covariance matrix.}
\item{nSamples}{The number of sets of missing data used to estimate the covariance matrix.}
}
\description{
Estimate the covariance matrix for a certain set of estimates. If one of the variances is negative (due to too much simulation noise), 
the algorithm continues to sample more sets of missing data until a certain maximum is reached. Note that difficulties estimating 
the covariance matrix could also indicate identifiability issues of the model. In this case, simplifying the model (e.g. lowering the number
of intervals in the piecewise constant function) might solve the problem.
}
\examples{
## See \link{StEM} function.
 
}
\author{
Michel H. Hof, \email{m.h.hof@amsterdamumc.nl}
}
